+++ entity/src/epilot/sdk.py
@@ -39,7 +39,7 @@ class Epilot:
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 epilot_auth: str,
+                 security: shared.Security = None,
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: dict[str, str] = None,
@@ -48,8 +48,8 @@ class Epilot:
                  ) -> None:
         """Instantiates the SDK configuring it with the provided parameters.
         
-        :param epilot_auth: The epilot_auth required for authentication
-        :type epilot_auth: str
+        :param security: The security details required for authentication
+        :type security: shared.Security
         :param server_idx: The index of the server to use for all operations
         :type server_idx: int
         :param server_url: The server URL to use for all operations